+++ server/api_friend.go
@@ -24,17 +24,22 @@ import (
 	"google.golang.org/grpc/status"
 )
 
-func (s *ApiServer) ListFriends(ctx context.Context, in *empty.Empty) (*api.Friends, error) {
+func (s *ApiServer) ListFriends(ctx context.Context, in *api.ListFriendsRequest) (*api.FriendList, error) {
 	userID := ctx.Value(ctxUserIDKey{}).(uuid.UUID)
 
 	// Before hook.
 	if fn := s.runtime.BeforeListFriends(); fn != nil {
 		beforeFn := func(clientIP, clientPort string) error {
-			err, code := fn(ctx, s.logger, userID.String(), ctx.Value(ctxUsernameKey{}).(string), ctx.Value(ctxExpiryKey{}).(int64), clientIP, clientPort)
+			result, err, code := fn(ctx, s.logger, userID.String(), ctx.Value(ctxUsernameKey{}).(string), ctx.Value(ctxExpiryKey{}).(int64), clientIP, clientPort, in)
 			if err != nil {
 				return status.Error(code, err.Error())
 			}
-			// Empty input never overridden.
+			if result == nil {
+				// If result is nil, requested resource is disabled.
+				s.logger.Warn("Intercepted a disabled resource.", zap.Any("resource", ctx.Value(ctxFullMethodKey{}).(string)), zap.String("uid", userID.String()))
+				return status.Error(codes.NotFound, "Requested resource was not found.")
+			}
+			in = result
 			return nil
 		}
 
@@ -45,8 +50,26 @@ func (s *ApiServer) ListFriends(ctx context.Context, in *empty.Empty) (*api.Frie
 		}
 	}
 
-	friends, err := GetFriends(ctx, s.logger, s.db, s.tracker, userID)
+	limit := 100
+	if in.GetLimit() != nil {
+		if in.GetLimit().Value < 1 || in.GetLimit().Value > 100 {
+			return nil, status.Error(codes.InvalidArgument, "Invalid limit - limit must be between 1 and 100.")
+		}
+		limit = int(in.GetLimit().Value)
+	}
+
+	state := in.GetState()
+	if state != nil {
+		if state := in.GetState().Value; state < 0 || state > 3 {
+			return nil, status.Error(codes.InvalidArgument, "Invalid state - state must be between 0 and 3.")
+		}
+	}
+
+	friends, err := GetFriends(ctx, s.logger, s.db, s.tracker, userID, limit, state, in.GetCursor())
 	if err != nil {
+		if err == ErrFriendInvalidCursor {
+			return nil, status.Error(codes.InvalidArgument, "Cursor is invalid.")
+		}
 		return nil, status.Error(codes.Internal, "Error while trying to list friends.")
 	}