+++ server/runtime_lua_nakama.go
@@ -5099,7 +5099,25 @@ func (n *RuntimeLuaNakamaModule) groupUsersList(l *lua.LState) int {
 		return 0
 	}
 
-	res, err := ListGroupUsers(l.Context(), n.logger, n.db, n.tracker, groupID)
+	limit := l.OptInt(2, 100)
+	if limit < 1 || limit > 100 {
+		l.ArgError(2, "expects limit to be 1-100")
+		return 0
+	}
+
+	state := l.OptInt(3, -1)
+	var stateWrapper *wrappers.Int32Value
+	if state != -1 {
+		if state < 0 || state > 3 {
+			l.ArgError(2, "expects state to be 0-3")
+			return 0
+		}
+		stateWrapper = &wrappers.Int32Value{Value: int32(state)}
+	}
+
+	cursor := l.OptString(4, "")
+
+	res, err := ListGroupUsers(l.Context(), n.logger, n.db, n.tracker, groupID, limit, stateWrapper, cursor)
 	if err != nil {
 		l.RaiseError("error while trying to list users in a group: %v", err.Error())
 		return 0
@@ -5151,7 +5169,12 @@ func (n *RuntimeLuaNakamaModule) groupUsersList(l *lua.LState) int {
 	}
 
 	l.Push(groupUsers)
-	return 1
+	if res.Cursor == "" {
+		l.Push(lua.LNil)
+	} else {
+		l.Push(lua.LString(res.Cursor))
+	}
+	return 2
 }
 
 func (n *RuntimeLuaNakamaModule) userGroupsList(l *lua.LState) int {
@@ -5161,7 +5184,25 @@ func (n *RuntimeLuaNakamaModule) userGroupsList(l *lua.LState) int {
 		return 0
 	}
 
-	res, err := ListUserGroups(l.Context(), n.logger, n.db, userID)
+	limit := l.OptInt(2, 100)
+	if limit < 1 || limit > 100 {
+		l.ArgError(2, "expects limit to be 1-100")
+		return 0
+	}
+
+	state := l.OptInt(3, -1)
+	var stateWrapper *wrappers.Int32Value
+	if state != -1 {
+		if state < 0 || state > 3 {
+			l.ArgError(2, "expects state to be 0-3")
+			return 0
+		}
+		stateWrapper = &wrappers.Int32Value{Value: int32(state)}
+	}
+
+	cursor := l.OptString(4, "")
+
+	res, err := ListUserGroups(l.Context(), n.logger, n.db, userID, limit, stateWrapper, cursor)
 	if err != nil {
 		l.RaiseError("error while trying to list groups for a user: %v", err.Error())
 		return 0
@@ -5201,7 +5242,12 @@ func (n *RuntimeLuaNakamaModule) userGroupsList(l *lua.LState) int {
 	}
 
 	l.Push(userGroups)
-	return 1
+	if res.Cursor == "" {
+		l.Push(lua.LNil)
+	} else {
+		l.Push(lua.LString(res.Cursor))
+	}
+	return 2
 }
 
 func (n *RuntimeLuaNakamaModule) accountUpdateId(l *lua.LState) int {